Adding 404 handler error

+++ controllers/categoryController.js
@@ -12,20 +12,30 @@ const categoryController = {
 
   },
   categoryDelete: (req, res) => {
-    db.Category.destroy({
-      where: {
-        id: req.params.id,
-      },
-    })
-      .then((result) => {
-        const resolve = {
-          status: 204,
-          message: 'Category deleted successfully!',
-          data: result,
-        };
-        res.json(resolve);
+    const category = db.findByPk(req.params.id);
+    if (category === null) {
+      const resolve = {
+        status: 404,
+        message: 'Category not found',
+        data: category,
+      };
+      res.json(resolve);
+    } else {
+      db.Category.destroy({
+        where: {
+          id: req.params.id,
+        },
       })
-      .catch((error) => res.json(error));
+        .then((result) => {
+          const resolve = {
+            status: 204,
+            message: 'Category deleted successfully!',
+            data: result,
+          };
+          res.json(resolve);
+        })
+        .catch((error) => res.json(error));
+    }
   },
   // End Categories CRUD
 };
